requirements.txt
flask
pytesseract
opencv-python
pillow
